+++ src/pages/home/components/HomeVideoSlider.tsx
@@ -6,36 +6,12 @@ import 'swiper/css';
 import 'swiper/css/pagination';
 import HomeVideo from './HomeVideo';
 import VideoInfo from './VideoInfo';
-import Skeleton from './Skeleton';
-
-export interface HomeVideoSliderProps {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data: any;
-  muted: boolean;
-  setMuted: (muted: boolean) => void;
-  setOpacity: (opacity: number) => void;
-  fetchNextPage: () => void;
-}
-
-interface ShortFormPage {
-  error: string;
-  response: {
-    hasNext: boolean;
-    nextPage: number;
-    shortForms: ShortForm[];
-  };
-  success: boolean;
-}
-interface ShortForm {
-  adoptionStatus: string;
-  age: string;
-  likeCount: string;
-  name: string;
-  petId: number;
-  profileShortFormUrl: string;
-  shelterId: number;
-  shelterName: string;
-}
+import Skeleton from './HomeSkeleton';
+import {
+  HomeVideoSliderProps,
+  ShortFormPage,
+  ShortFormVideo,
+} from '../homeType';
 
 const HomeVideoSlider = (props: HomeVideoSliderProps) => {
   const { data, muted, setMuted, setOpacity, fetchNextPage } = props;
@@ -85,13 +61,58 @@ const HomeVideoSlider = (props: HomeVideoSliderProps) => {
       keyboard={{ enabled: true, pageUpDown: true, onlyInViewport: true }}
     >
       {data?.pages.map((page: ShortFormPage, pagesIndex: number) =>
-        page.response.shortForms.map((shortForm: ShortForm, index: number) => {
-          if (pagesIndex * 5 + index === 5 * data.pages.length - 1) {
-            return (
-              <SwiperSlide key={shortForm.profileShortFormUrl + index}>
-                <div ref={nextPageRef}></div>
+        page.response.shortForms.map(
+          (shortForm: ShortFormVideo, index: number) => {
+            if (pagesIndex * 5 + index === 5 * data.pages.length - 1) {
+              return (
+                <SwiperSlide key={shortForm.profileShortFormUrl + index}>
+                  <div ref={nextPageRef}></div>
 
+                  <Swiper
+                    modules={[A11y]}
+                    grabCursor={true}
+                    autoHeight={true}
+                    direction={'horizontal'}
+                    onSlideNextTransitionEnd={() => {
+                      navigate(`/pet/${shortForm.petId}`);
+                    }}
+                  >
+                    <SwiperSlide>
+                      <div className="h-[70vh]">
+                        <HomeVideo
+                          url={shortForm.profileShortFormUrl}
+                          muted={muted}
+                          handleDoubleClick={handleDoubleClick}
+                          hovering={hovering}
+                          setHovering={setHovering}
+                          playing={playing}
+                          setPlaying={setPlaying}
+                          index={pagesIndex * 5 + index}
+                        />
+                      </div>
+                    </SwiperSlide>
+                    <SwiperSlide>
+                      <Skeleton text="동물 정보를 가져오는 중입니다🐶" />
+                    </SwiperSlide>
+                  </Swiper>
+                  <div className="w-full flex justify-center">
+                    <VideoInfo
+                      name={shortForm.name}
+                      adoptionStatus={shortForm.adoptionStatus}
+                      shelterName={shortForm.shelterName}
+                      age={shortForm.age}
+                    />
+                  </div>
+                </SwiperSlide>
+              );
+            }
+            return (
+              <SwiperSlide
+                key={shortForm.profileShortFormUrl + index}
+                className="flex justify-center items-centerr"
+              >
                 <Swiper
+                  className="bg-black"
                   modules={[A11y]}
                   grabCursor={true}
                   autoHeight={true}
@@ -101,7 +122,7 @@ const HomeVideoSlider = (props: HomeVideoSliderProps) => {
                   }}
                 >
                   <SwiperSlide>
-                    <div className="h-[70vh]">
+                    <div className="flex flex-col h-[70vh] items-center justify-center">
                       <HomeVideo
                         url={shortForm.profileShortFormUrl}
                         muted={muted}
@@ -128,51 +149,8 @@ const HomeVideoSlider = (props: HomeVideoSliderProps) => {
                 </div>
               </SwiperSlide>
             );
-          }
-          return (
-            <SwiperSlide
-              key={shortForm.profileShortFormUrl + index}
-              className="flex justify-center items-centerr"
-            >
-              <Swiper
-                className="bg-black"
-                modules={[A11y]}
-                grabCursor={true}
-                autoHeight={true}
-                direction={'horizontal'}
-                onSlideNextTransitionEnd={() => {
-                  navigate(`/pet/${shortForm.petId}`);
-                }}
-              >
-                <SwiperSlide>
-                  <div className="flex flex-col h-[70vh] items-center justify-center">
-                    <HomeVideo
-                      url={shortForm.profileShortFormUrl}
-                      muted={muted}
-                      handleDoubleClick={handleDoubleClick}
-                      hovering={hovering}
-                      setHovering={setHovering}
-                      playing={playing}
-                      setPlaying={setPlaying}
-                      index={pagesIndex * 5 + index}
-                    />
-                  </div>
-                </SwiperSlide>
-                <SwiperSlide>
-                  <Skeleton text="동물 정보를 가져오는 중입니다🐶" />
-                </SwiperSlide>
-              </Swiper>
-              <div className="w-full flex justify-center">
-                <VideoInfo
-                  name={shortForm.name}
-                  adoptionStatus={shortForm.adoptionStatus}
-                  shelterName={shortForm.shelterName}
-                  age={shortForm.age}
-                />
-              </div>
-            </SwiperSlide>
-          );
-        }),
+          },
+        ),
       )}
     </Swiper>
   );
