+++ src/pages/home/components/VideoDragBar.tsx
